NewMessage 분리
